Tests/LibWeb/Text/expected/canvas/global-composite-operation-porter-duff.txt

clear: Destination: rgb(0, 0, 0)
clear: Source: rgb(0, 0, 0)
clear: Combined: rgb(0, 0, 0)
copy: Destination: rgb(0, 0, 0)
copy: Source: rgb(100, 195, 230)
copy: Combined: rgb(100, 195, 230)
source-over: Destination: rgb(255, 220, 0)
source-over: Source: rgb(100, 195, 230)
source-over: Combined: rgb(155, 205, 155)
destination-over: Destination: rgb(255, 220, 0)
destination-over: Source: rgb(100, 195, 230)
destination-over: Combined: rgb(200, 210, 75)
source-in: Destination: rgb(0, 0, 0)
source-in: Source: rgb(0, 0, 0)
source-in: Combined: rgb(0, 0, 0)
destination-in: Destination: rgb(0, 0, 0)
destination-in: Source: rgb(0, 0, 0)
destination-in: Combined: rgb(0, 0, 0)
source-out: Destination: rgb(0, 0, 0)
source-out: Source: rgb(100, 195, 230)
source-out: Combined: rgb(100, 195, 230)
destination-out: Destination: rgb(0, 0, 0)
destination-out: Source: rgb(0, 0, 0)
destination-out: Combined: rgb(0, 0, 0)
source-atop: Destination: rgb(0, 0, 0)
source-atop: Source: rgb(0, 0, 0)
source-atop: Combined: rgb(0, 0, 0)
destination-atop: Destination: rgb(0, 0, 0)
destination-atop: Source: rgb(100, 195, 230)
destination-atop: Combined: rgb(100, 195, 230)
xor: Destination: rgb(255, 220, 0)
xor: Source: rgb(100, 195, 230)
xor: Combined: rgb(180, 210, 115)
lighter: Destination: rgb(255, 220, 0)
lighter: Source: rgb(100, 195, 230)
lighter: Combined: rgb(180, 210, 115)
